fix(modexp): e bit loop direction
We were looping the wrong direction over the e input. Used a bit of a
hacky fix by adding an extra state 'start' where we shift the bits of
the e-shiftreg until we reach the first one from the msb end before we
actually start the calculations. The bit scanner was removed, and the
input to the other shiftreg is simply a '1'.